[feat/margin_Firstpage_Calculate] 1. 마진 2 페이지 조회 구현, 2. CVR 계산을 위해 margin 테이블 전환 매출주문 수 추가, 3. 1개당 마진 Long 으로 수

+++ src/main/java/growup/spring/springserver/marginforcampaign/domain/MarginForCampaign.java
@@ -25,7 +25,7 @@ public class MarginForCampaign {
     private Long mfcSalePrice;
     private Long mfcTotalPrice; // 총비용
     private Long mfcCostPrice; // 원가
-    private Double mfcPerPiece; //  1개당 마진
+    private Long mfcPerPiece; //  1개당 마진
     private Double mfcZeroRoas; // 제로 Roas
 
     @ManyToOne
