+++ src/main/java/growup/spring/springserver/marginforcampaign/dto/MarginForCampaignResDto.java
@@ -11,6 +11,6 @@ public class MarginForCampaignResDto {
     private Long mfcSalePrice; // 판매가
     private Long mfcTotalPrice; // 총비용
     private Long mfcCostPrice; // 원가
-    private Double mfcPerPiece; //  1개당 마진
+    private Long mfcPerPiece; //  1개당 마진
     private Double mfcZeroRoas; // 제로 Roas
 }
