+++ src/main/java/growup/spring/springserver/marginforcampaign/dto/MfcDto.java
@@ -10,6 +10,6 @@ public class MfcDto {
     private Long mfcSalePrice;
     private Long mfcTotalPrice; // 총비용
     private Long mfcCostPrice; // 원가
-    private Double mfcPerPiece; //  1개당 마진
+    private Long mfcPerPiece; //  1개당 마진
     private Double mfcZeroRoas; // 제로 Roas
 }
